+++ ng2-rike/errors.component.d.ts
@@ -2,7 +2,7 @@ import { OnInit, OnDestroy } from "@angular/core";
 import { ErrorCollector } from "./error-collector";
 import { FieldErrors, FieldError } from "./field-error";
 export declare class RikeErrorsComponent implements OnInit, OnDestroy {
-    private _collector?;
+    private _collector;
     private _field?;
     private _errors;
     private _initialized;