+++ ng2-rike/event.js
@@ -1,4 +1,3 @@
-"use strict";
 var __extends = (this && this.__extends) || function (d, b) {
     for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
     function __() { this.constructor = d; }
@@ -11,18 +10,17 @@ var __extends = (this && this.__extends) || function (d, b) {
  *
  * Use [provideEventSource] function to register event sources.
  */
-var RikeEventSource = (function () {
+export var RikeEventSource = (function () {
     function RikeEventSource() {
     }
     return RikeEventSource;
 }());
-exports.RikeEventSource = RikeEventSource;
 /**
  * Basic REST-like resource access event.
  *
  * Such events are emitted by [Rike event sources][RikeEventsSource].
  */
-var RikeEvent = (function () {
+export var RikeEvent = (function () {
     function RikeEvent() {
     }
     Object.defineProperty(RikeEvent.prototype, "target", {
@@ -49,11 +47,10 @@ var RikeEvent = (function () {
     });
     return RikeEvent;
 }());
-exports.RikeEvent = RikeEvent;
 /**
  * An event emitted when operation on a REST-like resource is started.
  */
-var RikeOperationEvent = (function (_super) {
+export var RikeOperationEvent = (function (_super) {
     __extends(RikeOperationEvent, _super);
     function RikeOperationEvent(_operation) {
         _super.call(this);
@@ -103,11 +100,10 @@ var RikeOperationEvent = (function (_super) {
     });
     return RikeOperationEvent;
 }(RikeEvent));
-exports.RikeOperationEvent = RikeOperationEvent;
 /**
  * An event emitted when operation on a REST-like resource is successfully completed.
  */
-var RikeSuccessEvent = (function (_super) {
+export var RikeSuccessEvent = (function (_super) {
     __extends(RikeSuccessEvent, _super);
     function RikeSuccessEvent(_operation, _result) {
         _super.call(this);
@@ -158,13 +154,12 @@ var RikeSuccessEvent = (function (_super) {
     });
     return RikeSuccessEvent;
 }(RikeEvent));
-exports.RikeSuccessEvent = RikeSuccessEvent;
 /**
  * An event emitted when operation on a REST-like resource is failed.
  *
  * An object of this type is also reported as an error when some internal exception occurs.
  */
-var RikeErrorEvent = (function (_super) {
+export var RikeErrorEvent = (function (_super) {
     __extends(RikeErrorEvent, _super);
     function RikeErrorEvent(_operation, _error) {
         _super.call(this);
@@ -215,13 +210,12 @@ var RikeErrorEvent = (function (_super) {
     });
     return RikeErrorEvent;
 }(RikeEvent));
-exports.RikeErrorEvent = RikeErrorEvent;
 /**
  * An event emitted when operation on a REST-like resource caused an exception.
  *
  * An object of this type is reported as an error.
  */
-var RikeExceptionEvent = (function (_super) {
+export var RikeExceptionEvent = (function (_super) {
     __extends(RikeExceptionEvent, _super);
     function RikeExceptionEvent(operation, error, _errorResponse) {
         _super.call(this, operation, error);
@@ -236,11 +230,10 @@ var RikeExceptionEvent = (function (_super) {
     });
     return RikeExceptionEvent;
 }(RikeErrorEvent));
-exports.RikeExceptionEvent = RikeExceptionEvent;
 /**
  * An event emitted when operation on a REST-like resource returned error.
  */
-var RikeErrorResponseEvent = (function (_super) {
+export var RikeErrorResponseEvent = (function (_super) {
     __extends(RikeErrorResponseEvent, _super);
     function RikeErrorResponseEvent(operation, _errorResponse) {
         _super.call(this, operation, _errorResponse.error || _errorResponse);
@@ -255,11 +248,10 @@ var RikeErrorResponseEvent = (function (_super) {
     });
     return RikeErrorResponseEvent;
 }(RikeErrorEvent));
-exports.RikeErrorResponseEvent = RikeErrorResponseEvent;
 /**
  * An event emitted when operation on a REST-like resource is cancelled.
  */
-var RikeCancelEvent = (function (_super) {
+export var RikeCancelEvent = (function (_super) {
     __extends(RikeCancelEvent, _super);
     function RikeCancelEvent(operation, _cancelledBy) {
         _super.call(this, operation, _cancelledBy || "cancel");
@@ -288,6 +280,5 @@ var RikeCancelEvent = (function (_super) {
     });
     return RikeCancelEvent;
 }(RikeErrorEvent));
-exports.RikeCancelEvent = RikeCancelEvent;
 
 //# sourceMappingURL=event.js.map