+++ CHANGELOG.md
@@ -3,3 +3,15 @@
 ### Features
 
 * Allow to specify arrays of `StatusLabelMap` for status labels.
+
+
+# 0.1.2 (2016-09-03)
+
+### Features
+
+* *npm package:* Use ES2015 for primary build.
+* *bundle:* Add UMD bundle `bundles/ng2-rike.umd.js`.
+
+### Peer Dependencies
+
+* Angular2 rc6